Update jquery.keynav.js
Changed the $(document).keydown to a $(document).bind("keydown",...) with a namespace so that new instances of keynav on the same page do not interfere. This allows a dialog window to obtain a keynav without influence on the main page.

You do need to re-keynav the main page when the dialog is closed.